feat(transforms): convert $checkpoint using data_store prop
Get the [worfklow] $checkpoint value from the data store, async, at the
start of the action and set it at the end of the action. Use a
try...finally statement to ensure the value is set even if the action
throws an error.

For example:

- Input:

```js
$checkpoint ||= {}
$checkpoint["foo"] = "bar"
```

- Output:

```js
const $checkpoint = await this.db.get("$checkpoint")
try {
  $checkpoint ||= {}
  $checkpoint["foo"] = "bar"
} finally {
  await this.db.set("$checkpoint", $checkpoint)
}
```

Partially addresses #8